requirements.txt
absl-py==1.2.0
astunparse==1.6.3
bidict==0.22.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-SocketIO==4.3.1
flatbuffers==1.12
gast==0.4.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.28.1
six==1.16.0
tensorflow-cpu==2.9.1
termcolor==1.1.0
urllib3==1.26.10
Werkzeug==2.0.1
wrapt==1.14.1
